move check_server_port to util.py && lint
